+++ src/plugins/stake-plugins/generic/policyAdapters/types.ts
@@ -2,11 +2,12 @@ import { EdgeCurrencyWallet } from 'edge-core-js'
 
 import { AssetId, ChangeQuote, StakePolicy, StakePosition } from '../../types'
 import { CoreumNativeSkateKitAdapterConfig } from './CoreumStakeKitAdaptor'
+import { EthereumPooledKilnAdapterConfig } from './EthereumKilnAdaptor'
 import { GlifInfinityPoolAdapterConfig } from './GlifInfinityPoolAdapter'
 import { TarotPoolAdapterConfig } from './TarotPoolAdaptor'
 
 // Union of all adapter configs
-export type StakeAdapterConfig = GlifInfinityPoolAdapterConfig | TarotPoolAdapterConfig | CoreumNativeSkateKitAdapterConfig
+export type StakeAdapterConfig = CoreumNativeSkateKitAdapterConfig | EthereumPooledKilnAdapterConfig | GlifInfinityPoolAdapterConfig | TarotPoolAdapterConfig
 
 export interface StakePolicyAdapter {
   stakePolicyId: string